Modularize the script by separating some functions into the new files functions.py and terminal_colors.py
